Initial integration of synchronous sockets into WebDriver. Currently a clone of the async interface.
