+++ lib/html.dart
@@ -19,12 +19,12 @@ import 'dart:convert' show JSON;
 import 'dart:html' show HttpRequest, ProgressEvent;
 
 import 'package:webdriver/support/async.dart' show Lock;
-import 'package:webdriver/core.dart' as core
+import 'package:webdriver/async_core.dart' as core
     show createDriver, fromExistingSession, WebDriver;
-import 'package:webdriver/src/command_processor.dart' show CommandProcessor;
-import 'package:webdriver/src/exception.dart' show WebDriverException;
+import 'package:webdriver/src/async/command_processor.dart' show CommandProcessor;
+import 'package:webdriver/src/async/exception.dart' show WebDriverException;
 
-export 'package:webdriver/core.dart' hide createDriver, fromExistingSession;
+export 'package:webdriver/async_core.dart' hide createDriver, fromExistingSession;
 
 final Uri defaultUri = Uri.parse('http://127.0.0.1:4444/wd/hub/');
