+++ test/keyboard_test.dart
@@ -17,7 +17,7 @@ library webdriver.keyboard_test;
 import 'dart:io';
 
 import 'package:test/test.dart';
-import 'package:webdriver/core.dart';
+import 'package:webdriver/async_core.dart';
 
 import 'test_util.dart';
