+++ test/navigation_test.dart
@@ -16,7 +16,7 @@ library webdriver.navigation_test;
 
 import 'package:test/test.dart';
 import 'package:webdriver/support/async.dart';
-import 'package:webdriver/core.dart';
+import 'package:webdriver/async_core.dart';
 
 import 'test_util.dart';
